Tests/LibWeb/Text/expected/css/shadow-root-dirty-style-update-before-layout.txt

Slot color after forced layout: rgb(0, 0, 255)
